Add files via upload
This PDF contains the final written report for our undergraduate research replication of Burke et al. (2009), “Warming increases the risk of civil war in Africa.” The report details our methodology, replication process, results, and a discussion of discrepancies between the original study and our findings. Prepared as part of a student in the Data Science for Economists course at UC Berkeley.